Move part of spack.util.path to llnl.path
This moves the part that is needed by other llnl subpackages,
and doesn't have dependencies besides the Python std libs.
