+++ lib/spack/spack/test/directory_layout.py
@@ -12,11 +12,12 @@ from pathlib import Path
 
 import pytest
 
+from llnl.path import path_to_os_path
+
 import spack.paths
 import spack.repo
 from spack.directory_layout import DirectoryLayout, InvalidDirectoryLayoutParametersError
 from spack.spec import Spec
-from spack.util.path import path_to_os_path
 
 # number of packages to test (to reduce test time)
 max_packages = 10
