+++ lib/spack/spack/url.py
@@ -31,12 +31,12 @@ import pathlib
 import re
 
 import llnl.url
+from llnl.path import convert_to_posix_path
 from llnl.util.tty.color import cescape, colorize
 
 import spack.error
 import spack.util.web
 import spack.version
-from spack.util.path import convert_to_posix_path
 
 #
 # Note: We call the input to most of these functions a "path" but the functions
